Fix bug with initialization of the store and release notes
