+++ shell/components/nav/NotificationCenter/index.vue
@@ -15,15 +15,8 @@ import {
   RcDropdownTrigger
 } from '@components/RcDropdown';
 import { StoredNotification } from '@shell/types/notifications';
-import { deriveKey, encrypt, decrypt } from '@shell/utils/crypto/encryption';
-
-/**
- * Expire notifications in seconds (14 days)
- */
-const EXPIRY = 14 * 24 * 60 * 60;
-
-// Caches on load
-let encryptionKey: CryptoKey;
+import { encrypt } from '@shell/utils/crypto/encryption';
+import { loadFromString } from '@shell/utils/notifications';
 
 const store = useStore();
 const allNotifications = computed(() => store.getters['notifications/all']);
@@ -42,11 +35,12 @@ const open = (opened: boolean) => {
 };
 
 const localStorageKey = store.getters['notifications/localStorageKey'];
+const encryptionKey = store.getters['notifications/encryptionKey'];
 
 const localStorageEventHandler = async(ev: any) => {
   if (ev.key === localStorageKey) {
     try {
-      const data = await loadFromString(ev.newValue || '{}');
+      const data = await loadFromString(ev.newValue || '{}', encryptionKey);
 
       store.dispatch('notifications/load', data);
     } catch (e) {
@@ -58,9 +52,8 @@ const localStorageEventHandler = async(ev: any) => {
 /**
  * When notifications are updated, write to local storage
  */
-watch(allNotifications, async(newData) => {
+watch(allNotifications, async(newData: StoredNotification[]) => {
   try {
-    const localStorageKey = store.getters['notifications/localStorageKey'];
     const data = JSON.stringify(newData);
     const enc = await encrypt(data, encryptionKey);
 
@@ -70,87 +63,7 @@ watch(allNotifications, async(newData) => {
   }
 }, { deep: true });
 
-/**
- * Load data from a string
- *
- * @param dataString String containing JSON-serialized encrypted data
- */
-async function loadFromString(dataString: string) {
-  let notifications: StoredNotification[] = [];
-
-  try {
-    const data = JSON.parse(dataString || '{}');
-
-    // Legacy, not encrypted - perform basic check of data in the array
-    if (Array.isArray(data)) {
-      notifications = data.filter((n) => n.id && n.title);
-    }
-
-    if (data.cipher && data.iv) {
-      const decrypted = await decrypt(data, encryptionKey);
-      const persisted = JSON.parse(decrypted);
-
-      if (Array.isArray(persisted)) {
-        notifications = persisted as StoredNotification[];
-      } else {
-        console.error('Notification data is not in the expected format', persisted); // eslint-disable-line no-console
-      }
-    }
-  } catch (e) {
-    console.error('Unable to load notifications from data', e); // eslint-disable-line no-console
-  }
-
-  // Notifications needs to be an array, so check in case the data has been corrupted somehow and reset to empty state if so
-  // This is an extra check
-  if (!Array.isArray(notifications)) {
-    console.error('Notification data looks to be corrupt - ignoring persisted data'); // eslint-disable-line no-console
-
-    notifications = [] as StoredNotification[];
-  }
-
-  return notifications;
-}
-
-/**
- * Load the notifications from local storage on start up
- */
 onMounted(async() => {
-  const uuid = store.getters['notifications/userId'];
-  let notifications: StoredNotification[] = [];
-
-  try {
-    encryptionKey = await deriveKey(uuid);
-  } catch (e) {
-    console.error('Unable to generate encryption key for notifications', e); // eslint-disable-line no-console
-
-    return;
-  }
-
-  try {
-    const data = window.localStorage.getItem(localStorageKey) || '{}';
-
-    notifications = await loadFromString(data);
-  } catch (e) {
-    console.error('Unable to read notifications from local storage', e); // eslint-disable-line no-console
-  }
-
-  // Expire old notifications
-  const now = new Date();
-
-  notifications = notifications.filter((n: StoredNotification) => {
-    // Try ... catch in case the date parsing fails
-    try {
-      const created = new Date(n.created);
-      const diff = (now.getTime() - created.getTime()) / 1000; // Diff in seconds
-
-      return diff < EXPIRY;
-    } catch (e) {}
-
-    return true;
-  });
-
-  await store.dispatch('notifications/load', notifications);
-
   // Listen to storage events, so if the UI is open in multiple tabs, notifications in one tab will be sync'ed across all tabs
   window.addEventListener('storage', localStorageEventHandler);
 });
